live_orchestrator/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
requests==2.32.3
python-dotenv==1.0.1
# For real Gemini integration (optional in this scaffold):
google-genai>=1.5.0
